tag2domain_api/requirements.txt
attrs==19.3.0
click==7.1.2
envelope==1.2.4
fastapi==0.60.1
h11==0.9.0
httptools==0.1.1
importlib-metadata==1.7.0
jsonpickle==1.4.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary>=2.8
py==1.9.0
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.0.0
python-gnupg==0.4.6
python-magic==0.4.18
six==1.15.0
starlette==0.13.6
uvicorn==0.11.7
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
zipp==3.1.0
python-dotenv==0.14.0
dnspython==1.16.0
filelock==3.0.12
idna==2.10
iniconfig==1.0.0
py3-validate-email==0.2.9
toml==0.10.1
orjson>=3.4.6
pyyaml>=5.4.1